refactor: 낙관적 락에서 비관적 락으로 변경
