+++ backend/src/main/java/com/now/naaga/like/presentation/PlaceLikeController.java
@@ -1,7 +1,6 @@
 package com.now.naaga.like.presentation;
 
 import com.now.naaga.auth.presentation.annotation.Auth;
-import com.now.naaga.like.application.PlaceLikeFacade;
 import com.now.naaga.like.application.PlaceLikeService;
 import com.now.naaga.like.application.dto.ApplyLikeCommand;
 import com.now.naaga.like.application.dto.CancelLikeCommand;
@@ -29,13 +28,9 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 public class PlaceLikeController {
 
-    private final PlaceLikeFacade placeLikeFacade;
-
     private final PlaceLikeService placeLikeService;
 
-    public PlaceLikeController(final PlaceLikeFacade placeLikeFacade,
-                               final PlaceLikeService placeLikeService) {
-        this.placeLikeFacade = placeLikeFacade;
+    public PlaceLikeController(final PlaceLikeService placeLikeService) {
         this.placeLikeService = placeLikeService;
     }
 
@@ -46,7 +41,7 @@ public class PlaceLikeController {
         final ApplyLikeCommand command = ApplyLikeCommand.of(playerRequest,
                                                              placeId,
                                                              applyPlaceLikeRequest);
-        final PlaceLike placeLike = placeLikeFacade.applyLike(command);
+        final PlaceLike placeLike = placeLikeService.applyLike(command);
         final PlaceLikeResponse response = PlaceLikeResponse.from(placeLike);
         return ResponseEntity
                 .status(HttpStatus.CREATED)
@@ -58,7 +53,7 @@ public class PlaceLikeController {
     public ResponseEntity<Void> cancelPlaceLike(@Auth PlayerRequest playerRequest,
                                                 @PathVariable Long placeId) {
         final CancelLikeCommand cancelLikeCommand = CancelLikeCommand.of(playerRequest, placeId);
-        placeLikeFacade.cancelLike(cancelLikeCommand);
+        placeLikeService.cancelLike(cancelLikeCommand);
         return ResponseEntity
                 .status(HttpStatus.NO_CONTENT)
                 .build();
