+++ backend/src/main/java/com/now/naaga/place/application/PlaceStatisticsService.java
@@ -41,7 +41,7 @@ public class PlaceStatisticsService {
 
     public void plusLike(final PlusLikeCommand plusLikeCommand) {
         final Long placeId = plusLikeCommand.placeId();
-        final PlaceStatistics placeStatistics = placeStatisticsRepository.findByPlaceIdWithOptimisticLock(placeId)
+        final PlaceStatistics placeStatistics = placeStatisticsRepository.findByPlaceIdForUpdate(placeId)
                                                                          .orElseThrow(() -> new PlaceException(PlaceExceptionType.NO_EXIST));
         placeStatistics.plusLike();
     }
@@ -49,9 +49,8 @@ public class PlaceStatisticsService {
     public void subtractLike(final SubtractLikeCommand subtractLikeCommand) {
         final Long placeId = subtractLikeCommand.placeId();
 
-        final PlaceStatistics placeStatistics = placeStatisticsRepository.findByPlaceIdWithOptimisticLock(placeId)
+        final PlaceStatistics placeStatistics = placeStatisticsRepository.findByPlaceIdForUpdate(placeId)
                                                                          .orElseThrow(() -> new PlaceStatisticsException(PlaceStatisticsExceptionType.NOT_FOUND));
-
         placeStatistics.subtractLike();
     }
